third_party/blink/web_tests/platform/generic/media/media-fragments/TC0090-expected.txt


Title: TC0090
Fragment: '&&t=3,7'
Comment: Sprinkling & is OK.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
